TEST_ONLY Move test classes under javax.jmdns.impl package.
