+++ src/test/java/javax/jmdns/impl/ServiceInfoTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package javax.jmdns.test;
+package javax.jmdns.impl;
 
 import org.junit.jupiter.api.Test;
